Support project flavors for the android-library plugin
When this code was originally written, flavors were not supported by the android-library plugin. As such, when building a variation name, an optimization was put in place to skip the flavor as a piece of the ultimate variation name in the case of the android-library plugin. Now that flavors are supported for libraries, this optimization actually breaks the roboelectric plugin with an error such as "Error:Cannot add a SourceSet with name 'testDebug' as a SourceSet with that name already exists.". This is because a library today that *does* have flavors will cause this plugin to attempt to add the same-named test sourceset for each flavor+buildtype.